+++ pkg/cmd/grafana-cli/commands/cuetsify_command.go
@@ -45,12 +45,10 @@ var skipPaths = []string{
 	"public/app/plugins/panel/gauge/models.cue",
 	"public/app/plugins/panel/histogram/models.cue",
 	"public/app/plugins/panel/heatmap-new/models.cue",
-	"public/app/plugins/panel/stat/models.cue",
 	"public/app/plugins/panel/candlestick/models.cue",
 	"public/app/plugins/panel/state-timeline/models.cue",
 	"public/app/plugins/panel/status-history/models.cue",
 	"public/app/plugins/panel/table/models.cue",
-	"public/app/plugins/panel/text/models.cue",
 	"public/app/plugins/panel/timeseries/models.cue",
 	// All the cue files in this dir have to be individually excluded, even
 	// though the generator currently smooshes them all together