+++ public/app/plugins/panel/heatmap-new/models.gen.ts
@@ -8,10 +8,10 @@ import { HeatmapCalculationOptions } from 'app/features/transformers/calculateHe
 
 export const modelVersion = Object.freeze([1, 0]);
 
-export enum HeatmapSourceMode {
-  Auto = 'auto',
+export enum HeatmapMode {
+  Aggregated = 'agg',
   Calculate = 'calculate',
-  Data = 'data', // Use the data as is
+  Accumulated = 'acc', // accumulated
 }
 
 export enum HeatmapColorMode {
@@ -33,7 +33,6 @@ export interface HeatmapColorOptions {
   steps: number; // 2-128
 
   // Clamp the colors to the value range
-  field?: string;
   min?: number;
   max?: number;
 }
@@ -46,11 +45,15 @@ export interface HeatmapLegend {
   show: boolean;
 }
 
+export interface ExemplarConfig {
+  color: string;
+}
+
 export interface PanelOptions {
-  source: HeatmapSourceMode;
+  mode: HeatmapMode;
 
   color: HeatmapColorOptions;
-  heatmap?: HeatmapCalculationOptions;
+  calculate?: HeatmapCalculationOptions;
   showValue: VisibilityMode;
 
   cellGap?: number; // was cardPadding
@@ -62,10 +65,11 @@ export interface PanelOptions {
   legend: HeatmapLegend;
 
   tooltip: HeatmapTooltip;
+  exemplars: ExemplarConfig;
 }
 
 export const defaultPanelOptions: PanelOptions = {
-  source: HeatmapSourceMode.Auto,
+  mode: HeatmapMode.Aggregated,
   color: {
     mode: HeatmapColorMode.Scheme,
     scheme: 'Oranges',
@@ -81,6 +85,9 @@ export const defaultPanelOptions: PanelOptions = {
   },
   legend: {
     show: true,
+  },
+  exemplars: {
+    color: 'rgba(255,0,255,0.7)',
   },
   cellGap: 1,
 };